Allow external calling context to inject values.
To allow non-core projects to use devstack-gate to test their changes without
needing to modify devstack-gate itself, allow injection of values into
PROJECTS and ENABLED_SERVICES from the calling script. Also let the calling
script drop in a localrc fragment and tell devstack gate to not delete it.
That way a jenkins job could set those values and then call an unmodified
devstack-gate.

+++ devstack-vm-gate.sh
@@ -23,9 +23,20 @@ set -o errexit
 
 cd $BASE/new/devstack
 
-rm -f localrc
+# Allow calling context to pre-populate the localrc file with additional values
+if [ -z $KEEP_LOCALRC ] ; then
+    rm -f localrc
+fi
 
-ENABLED_SERVICES=g-api,g-reg,key,n-api,n-crt,n-obj,n-cpu,n-sch,horizon,mysql,rabbit,sysstat
+
+DEFAULT_ENABLED_SERVICES=g-api,g-reg,key,n-api,n-crt,n-obj,n-cpu,n-sch,horizon,mysql,rabbit,sysstat
+
+# Allow optional injection of ENABLED_SERVICES from the calling context
+if [ -z $ENABLED_SERVICES ] ; then
+    ENABLED_SERVICES=$DEFAULT_ENABLED_SERVICES
+else
+    ENABLED_SERVICES=$DEFAULT_ENABLED_SERVICES,$ENABLED_SERVICES
+fi
 
 if [ "$DEVSTACK_GATE_TEMPEST" -eq "1" ]; then
     ENABLED_SERVICES=$ENABLED_SERVICES,tempest
